services/voice/requirements.txt
faster-whisper
piper-tts
piper
openwakeword
sounddevice
numpy
httpx
keyboard
python-dotenv
pytest
pytest-mock
